requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
isort==4.3.20
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
Pillow==6.0.0
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.1
